add a git hook to use enable erb'd commit templates
This adds a `prepare-commit-msg` hook, which is run before your editor
opens your commit message, but after git has identified and loaded
your commit template (if any).

This requires ruby, so it _should_ be added to the `install` script,
though that can honestly be improved a bunch on its own.